Test 1
I have some questions about this page. I know is for copy and have different kind of codes, That I maybe use in the future, but in the end I don´t know how it works. Anyway this is my lesson for today, Tomorrow I´ll sumbit the next lesson.

Homework done!! .... BYE.